\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print}
\alias{print,moa-method}
\alias{print,moa.sup-method}
\alias{print,mgsa-method}
\title{ Methods for function \code{print} }
\description{
  Methods for function \code{print}
}

\section{Methods}{
\describe{

  \item{\code{signature(object = "moa")}}{
	  print "moa" class
  }

  \item{\code{signature(object = "moa.sup")}}{
    print "sup.moa" class
  }

  \item{\code{signature(object = "mgsa")}}{
    print "mgsa" class
  }
}}
\keyword{ methods }
\keyword{ moa-class }
\keyword{ sup.moa-class }
\keyword{ mgsa-class }
