\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show}
\alias{show,moa-method}
\alias{show,moa.sup-method}
\alias{show,mgsa-method}
\title{ Methods for function \code{show} }
\description{
  Methods for function \code{show}
}

\section{Methods}{
\describe{

  \item{\code{signature(object = "moa")}}{
	  show "moa" class
  }

  \item{\code{signature(object = "moa.sup")}}{
    show "sup.moa" class
  }

  \item{\code{signature(object = "mgsa")}}{
    show "mgsa" class
  }
}}
\keyword{ methods }
\keyword{ moa-class }
\keyword{ sup.moa-class }
\keyword{ mgsa-class }
