\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary}
\alias{summary,moa-method}
\alias{summary,moa.sup-method}
\alias{summary,mgsa-method}
\title{ Methods for function \code{summary} }
\description{
  Methods for function \code{summary}
}

\section{Methods}{
\describe{

  \item{\code{signature(object = "moa")}}{
	  summary "moa" class
  }

  \item{\code{signature(object = "moa.sup")}}{
    summary "sup.moa" class
  }

  \item{\code{signature(object = "mgsa")}}{
    summary "mgsa" class
  }
}}
\keyword{ methods }
\keyword{ moa-class }
\keyword{ sup.moa-class }
\keyword{ mgsa-class }
